+++ opensquirrel/passes/exporter/quantify_scheduler_exporter.py
@@ -15,6 +15,7 @@ from opensquirrel.ir import (
     Barrier,
     BlochSphereRotation,
     BsrAngleParam,
+    BsrFullParams,
     BsrNoParams,
     ControlledGate,
     CRk,
@@ -155,6 +156,9 @@ class _ScheduleCreator(IRVisitor):
         raise UnsupportedGateError(gate)
 
     def visit_bsr_no_params(self, gate: BsrNoParams) -> None:
+        self.visit_bloch_sphere_rotation(gate)
+
+    def visit_bsr_full_params(self, gate: BsrFullParams) -> None:
         self.visit_bloch_sphere_rotation(gate)
 
     def visit_bsr_angle_param(self, gate: BsrAngleParam) -> None: