+++ test/test_integration.py
@@ -17,7 +17,7 @@ from opensquirrel.passes.decomposer import (
 )
 from opensquirrel.passes.exporter import ExportFormat
 from opensquirrel.passes.merger import SingleQubitGatesMerger
-from opensquirrel.passes.validator import NativeGateValidator, RoutingValidator
+from opensquirrel.passes.validator import PrimitiveGateValidator, RoutingValidator
 
 
 def test_spin2plus_backend() -> None:
@@ -82,7 +82,7 @@ def test_spin2plus_backend() -> None:
     """
     connectivity = {"0": [1], "1": [0]}
 
-    native_gate_set = ["I", "X90", "mX90", "Y90", "mY90", "Rz", "CZ"]
+    primitive_gate_set = ["I", "X90", "mX90", "Y90", "mY90", "Rz", "CZ", "measure", "wait", "init", "barrier"]
 
     # Validate that the interactions in the circuit are possible given the chip topology
     qc.validate(validator=RoutingValidator(connectivity))
@@ -99,11 +99,11 @@ def test_spin2plus_backend() -> None:
     # Merge single-qubit gates
     qc.merge(merger=SingleQubitGatesMerger())
 
-    # Decompose single-qubit gates to spin backend native gates with McKay decomposer
+    # Decompose single-qubit gates to primitive gates with McKay decomposer
     qc.decompose(decomposer=McKayDecomposer())
 
-    # Validate that the compiled circuit is composed of gates that are in the native gate set
-    qc.validate(validator=NativeGateValidator(native_gate_set))
+    # Validate that the compiled circuit is composed of gates that are in the primitive gate set
+    qc.validate(validator=PrimitiveGateValidator(primitive_gate_set))
 
     assert (
         str(qc)
@@ -253,7 +253,7 @@ def test_hectoqubit_backend() -> None:
         "3": [2],
         "4": [2],
     }
-    native_gate_set = [
+    primitive_gate_set = [
         "I",
         "Rx",
         "X",
@@ -272,6 +272,8 @@ def test_hectoqubit_backend() -> None:
         "CZ",
         "measure",
         "reset",
+        "wait",
+        "init",
         "barrier",
     ]
 
@@ -290,11 +292,11 @@ def test_hectoqubit_backend() -> None:
     # Merge single-qubit gates
     qc.merge(merger=SingleQubitGatesMerger())
 
-    # Decompose single-qubit gates to HectoQubit backend native gates with the XYX decomposer
+    # Decompose single-qubit gates to primitive gates with the XYX decomposer
     qc.decompose(decomposer=ZXZDecomposer())
 
-    # Validate that the compiled circuit is composed of gates that are in the native gate set
-    qc.validate(validator=NativeGateValidator(native_gate_set))
+    # Validate that the compiled circuit is composed of gates that are in the primitive gate set
+    qc.validate(validator=PrimitiveGateValidator(primitive_gate_set))
 
     if importlib.util.find_spec("quantify_scheduler") is None:
         with pytest.raises(
@@ -735,7 +737,7 @@ def test_starmon7_backend() -> None:
         "6": [3, 4],
     }
 
-    native_gate_set = [
+    primitive_gate_set = [
         "I",
         "H",
         "X",
@@ -757,13 +759,17 @@ def test_starmon7_backend() -> None:
         "CR",
         "CRk",
         "SWAP",
+        "measure",
+        "wait",
+        "init",
+        "barrier",
     ]
 
     # Validate that the interactions in the circuit are possible given the chip topology
     qc.validate(validator=RoutingValidator(connectivity))
 
-    # Validate that the compiled circuit is composed of gates that are in the native gate set
-    qc.validate(validator=NativeGateValidator(native_gate_set))
+    # Validate that the compiled circuit is composed of gates that are in the primitive gate set
+    qc.validate(validator=PrimitiveGateValidator(primitive_gate_set))
 
     exported_circuit = qc.export(fmt=ExportFormat.CQASM_V1)