+++ CHANGELOG.md
@@ -14,14 +14,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `Rn` unitary instruction
+- `CZDecomposer` decomposer pass
 - `ShortestPathRouter` router pass
 - `RandomMapper` mapper pass
+- `AStarRouter` router pass
 
 ### Changed
 
 - libQASM 1.1.0 integrated (updated from 0.6.9)
 - Refactor: removed generators
 - Changed the `RoutingChecker` pass to a `RoutingValidator` pass
+- Changed use of `native` to `primitive`, e.g. `NativeGateValidator` is now `PrimitiveGateValidator`
 
 ## [0.3.1] - [ 2025-01-31 ]